examples/integrations/stripe/python/requirements.txt
stripe 
browserbasehq
playwright 
python-dotenv